+++ src/structure/mod.rs
@@ -1,3 +1,4 @@
 pub mod matrix;
 pub mod vector;
 pub mod polynomial;
+pub mod dual;